Remove duplicated request current url
