+++ main.jsx
@@ -9,13 +9,14 @@ var
     getScripts = require('./js/getScripts'),
     $ = require('jquery'),
     Showdown = require('./showdown.min'),
+    converter = new Showdown.converter(),
     getFinalJSON = require('./js/storages/getFinalJSON'),
-	getParaGraph = require('./js/storages/getParaGraph'),
-	finalMetionsArray = require('./js/storages/finalMetionsArray'),
-	finalMenuJsonArray = require('./js/storages/finalMenuJsonArray'),
-	finalListJsonArray = require('./js/storages/finalListJsonArray'),
+    getParaGraph = require('./js/storages/getParaGraph'),
+    finalMetionsArray = require('./js/storages/finalMetionsArray'),
+    finalMenuJsonArray = require('./js/storages/finalMenuJsonArray'),
+    finalListJsonArray = require('./js/storages/finalListJsonArray'),
 
-	CreateArticleList = require('./js/components/CreateArticleList'),
+    CreateArticleList = require('./js/components/CreateArticleList'),
 
     importUrl = require('./js/global').importUrl,
     cssUrl = require('./js/global').cssUrl,
@@ -32,69 +33,12 @@ var
     storeageKeys = [],
     href = window.location.href,
     is_list = false,
-    is_airticlelist = false,
+    isAirticlelist = require('./js/global').isAirticlelist,
     is_hashUrl = false,
-    is_cover = false;
+    is_cover = false,
+    addBootstrapLink = require('./js/addBootstrapLink');
 
- (function(){
-     'use strict';
-   
-    $('body').on('click','#myoffcanvas',function() {
-     $('.row-offcanvas').toggleClass('active');
-       $('.row-offcanvas-menu').toggleClass('active');
-    });
-   
-   // for plus tab
-    $('body').on('click','.plusIcon',function() {
-      $('.active').removeClass('active');
-      $('.removeParent').removeClass('collapsed');
-      $('.parentNodeA').removeClass('collapsed');
-      $('.new').addClass('active');
-      defaultList();
-    });
-    // for plus tab
-    
-    
-    // right menu click
-     $('body').on('click','.listView',function() {
-        var url=$(this).attr('data-url');
-          $('.listView').removeClass('active');
-      $(".listView").parent().removeClass("active");
-        $(this).addClass('active');
-      $(".active").parent().addClass("active");
-     getListJson(url);
-      });
-    
-    $('body').on('click','.articleView',function() {
-        var url=$(this).attr('data-url');
-          $(".listView").parent().removeClass("active");
-      $('.itemListWrp').css('display','none');
-      $("section").parent("article").css('display','block');
-          $('#titter_frame_container').css('display','block');      
-      
-      });
-    // right menu click
-   
- var addBootstrapLink = function(){
-         var link = document.createElement('link');
-         link.rel = 'stylesheet';
-         link.href = 'https://maxcdn.bootstrapcdn.com/bootstrap/3.3.1/css/bootstrap.min.css';
-         document.head.appendChild(link);
-
-         link = document.createElement('link');
-         link.rel = 'stylesheet';
-         link.href = cssUrl + theme + '/css/webrunes.css';
-         document.head.appendChild(link);
-
-         link = document.createElement('link');
-         link.rel = 'shortcut icon';
-         link.href = cssUrl + theme + '/ico/favicon.ico';
-         document.head.appendChild(link);
-     };
-   addBootstrapLink();
- })();
-
-var converter = new Showdown.converter();
+addBootstrapLink();
 
 var finalJson;
 var itemListArray = [];
@@ -103,7 +47,7 @@ var itemListArray = [];
 var CreateDomLeft = React.createClass({
   render: function() {    
     return (
-      <div className="col-xs-12 col-sm-3 col-md-2"><div className="navbar navbar-inverse main navbar-fixed-top row-offcanvas-menu"><div className="navbar-header tooltip-demo" id="topMenu"><ul className="nav menu pull-right"><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Call IA"><a className="btn btn-link btn-sm" href="#"><span className="glyphicon glyphicon-comment"></span></a></li><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Logout"><a className="btn btn-link btn-sm" href="#"><span className="glyphicon glyphicon-lock"></span></a></li><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Full screen"><a className="btn btn-link btn-sm" href="#"><span className="glyphicon glyphicon-fullscreen"></span></a></li><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Open/close menu"><a data-target=".navbar-collapse" data-toggle="collapse" className="btn btn-link btn-sm visible-xs collapsed" href="#"><span className="glyphicon glyphicon-align-justify"></span></a></li><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Show/hide the sidebar"><a data-toggle="offcanvas" id="myoffcanvas"  className="btn btn-link btn-sm visible-xs" href="#"><span className="glyphicon glyphicon-transfer"></span></a></li></ul><a title="" data-placement="right" data-toggle="tooltip" className="navbar-brand" href="webrunes-contact.htm" data-original-title="Contact us">{'&nbsp;'}</a></div><div className="navbar-collapse in"><div className="navbar-header" id="leftMenuwrp"><CreateLeftCommentMenus /></div></div></div></div>
+      <div className="col-xs-12 col-sm-3 col-md-2"><div className="navbar navbar-inverse main navbar-fixed-top row-offcanvas-menu"><div className="navbar-header tooltip-demo" id="topMenu"><ul className="nav menu pull-right"><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Call IA"><a className="btn btn-link btn-sm" href="#"><span className="glyphicon glyphicon-comment"></span></a></li><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Logout"><a className="btn btn-link btn-sm" href="#"><span className="glyphicon glyphicon-lock"></span></a></li><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Full screen"><a className="btn btn-link btn-sm" href="#"><span className="glyphicon glyphicon-fullscreen"></span></a></li><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Open/close menu"><a data-target=".navbar-collapse" data-toggle="collapse" className="btn btn-link btn-sm visible-xs collapsed" href="#"><span className="glyphicon glyphicon-align-justify"></span></a></li><li title="" data-placement="bottom" data-toggle="tooltip" data-original-title="Show/hide the sidebar"><a data-toggle="offcanvas" id="myoffcanvas"  className="btn btn-link btn-sm visible-xs" href="#"><span className="glyphicon glyphicon-transfer"></span></a></li></ul><a title="" data-placement="right" data-toggle="tooltip" className="navbar-brand" href="webrunes-contact.htm" data-original-title="Contact us">&nbsp;</a></div><div className="navbar-collapse in"><div className="navbar-header" id="leftMenuwrp"><CreateLeftCommentMenus /></div></div></div></div>
     );
   }
 });
@@ -125,48 +69,27 @@ var CreateLeftCommentMenus = React.createClass({
     <CreateLeftItemMenu />    
     );
   }           
-}); 
+});
 
 var CreateLeftItemMenu = React.createClass({
-  render: function() {
-      return (    
-    <section />
-      );
-  }
-}); 
-
+    render: function() {
+        return (    
+            <section />
+        );
+    }
+});
 
 var CreateDomRight = React.createClass({
   render: function() {
     return (
       <div className="col-xs-6 col-sm-4 col-md-3 sidebar-offcanvas" id="sidebar">
         <div className="sidebar-margin">
-          <CreateArticleMenus />
+          <CreateArticleItemMenu data={this.props.data}/>
         </div>
       </div>
     );
   }
 });
-
-// for right menu
-var CreateArticleMenus = React.createClass({
-  getInitialState: function() {
-    return {data: []};
-  },
-  componentDidMount: function() {
-  if(is_airticlelist==false){
-    getListJson(href); 
-  }else{
-     this.setState({data: finalMenuJsonArray});
-    }
-  },
-  render: function() {
-    return (
-            <CreateArticleItemMenu data={this.state.data} />
-    );
-  }
-});
-
 
 var CreateArticleItemMenu = React.createClass({
   render: function() {
@@ -203,7 +126,7 @@ var CreateCommentMenus = React.createClass({
   },
   render: function() {
     return (
-            <CreateItemMenu data={this.state.data} />
+        <CreateItemMenu data={this.state.data} />
     );
   }
 });
@@ -214,7 +137,9 @@ var CreateItemMenu = React.createClass({
       var commentMenustring = comment.articlename.replace(/\s/g, '_');
       var href = comment.url ? comment.url : '#' + commentMenustring;
       return (
-          <li key={index}><a href={href}>{comment.articlename}</a></li>   
+        <li key={index}>
+            <a href={href}>{comment.articlename}</a>
+        </li>   
       );
     });        
     
@@ -249,9 +174,9 @@ var CreateBreadcrumb = React.createClass({
          var htmlBreadcrumb='<ul class="breadcrumb controls tooltip-demo"><li title="Read time" data-placement="top" data-toggle="tooltip"><span class="glyphicon glyphicon-time"></span>4-5 minutes</li><li title="Last modified" data-placement="top" data-toggle="tooltip"><span class="glyphicon glyphicon-calendar"></span>30 May, 2014</li></ul><ul itemprop="breadcrumb" class="breadcrumb"><li class="active">Read</li><li><a href="#">Edit</a></li></ul>';
    
       var rawMarkup = converter.makeHtml(htmlBreadcrumb.toString());
-    if(rawMarkup=="") return false;
-      return (                    
-    <section dangerouslySetInnerHTML={{__html: rawMarkup}}></section>
+    if(rawMarkup === "") return null;
+      return (
+        <section dangerouslySetInnerHTML={{__html: rawMarkup}}></section>
       );
   }
 });
@@ -297,15 +222,14 @@ var Main = React.createClass({
   render: function() {
     window.complete_script = getScripts();
     finalJson = getFinalJSON(complete_script);
-    checkUrl(); // for check # url
     return (
-		<div id="content" className="container-liquid">
-			<div className="row row-offcanvas row-offcanvas-right">
-				<CreateDomLeft />
-				<CreateDomCenter />
-				<CreateDomRight />
-			</div>
-		</div>
+        <div id="content" className="container-liquid">
+            <div className="row row-offcanvas row-offcanvas-right">
+                <CreateDomLeft />
+                <CreateDomCenter />
+                <CreateDomRight data={finalMenuJsonArray} />
+            </div>
+        </div>
     );
   }
 });
@@ -419,213 +343,9 @@ var CommentForm = React.createClass({
   }
 });
 
-
-// get default List on click of plus tab
-function defaultList(){
-     plusArray= JSON.parse(localStorage.getItem('plusTabItem'));
-      //var url = importUrl + 'Default-WRIO-Theme/widget/itemList.htm';
-    
-    var url= themeImportUrl + 'itemList.htm';
-    
-    
-  if ( !$('.plusActive').hasClass('plusList')) {  // for check default list available or not
-    $.ajax({
-      url: url,
-      dataType: 'html',
-      success: function(data) {  
-        if(plusArray !== undefined && typeof plusArray[0] === 'object'){ 
-              $('#centerWrp').html(
-             data
-            .replace("{title}", plusArray[0].name)
-            .replace("{sub_title}",plusArray[0].name)
-            .replace("{about}",plusArray[0].about)
-            .replace("{image}", plusArray[0].image)
-            .replace("{url}", plusArray[0].url)
-            .replace("{created_date}","22 Jun 2013")
-            .replace("{rating}", "244")
-            .replace("{readers}", "1,634")
-            .replace("{access}", "Free")
-          );
-          $('#plusWrp').addClass('plusActive');
-          }
-      }
-    });   
-  }
-}
- 
- 
- // for list
-function getListJson(listurl){
-
-       var url=listurl;
-     if(url.indexOf("?cover") != -1){
-        is_cover=true;
-     }
-     var jsonItemArray = [];
-    $.get(url, function(result){
-           var rHtml = result.replace('<script type="text/javascript" src="http://wrio.s3-website-us-east-1.amazonaws.com/WRIO-InternetOS/WRIO.js"></script>', '');
-     
-     jQuery('<div/>', {
-        id: 'foo1',
-        css:{
-          display:'none'
-        },
-        html: rHtml
-      }).appendTo('body');
-           defaultPlusList=$("#foo1 [type^='application/ld+json']" ).html();
-       $('#foo1').remove();
-         defaultPlusListJson = JSON.parse(defaultPlusList);   
-       itemListArray=defaultPlusListJson.itemListElement;
-       getlist(itemListArray);
-       
-       if(is_list==true){
-        //  var htmlList=getlist();
-       }
-    }); 
-}
- 
- 
-// get List  for list 
-function getlist(itemListArray){
-      
-              var url = themeImportUrl + 'itemList.htm'; 
-        var coverHtml='<div class="{status}"><div style="background: url({contentUrl}) center center" class="img"></div><div class="carousel-caption"><div class="carousel-text"><h2>{name}</h2><ul class="features">{text}</ul></div></div></div>';
-         var contentdiv='<li><span class="glyphicon glyphicon-ok"></span>{content}</li>';         
-         var coverHeader='<ul class="breadcrumb"><li class="active">Cover</li></ul><div data-ride="carousel" class="carousel slide" id="cover-carousel"><div class="carousel-inner">';
-         var coverFooter='</div><a data-slide="prev" href="#cover-carousel" class="left carousel-control"><span class="glyphicon glyphicon-chevron-left"></span></a><a data-slide="next" href="#cover-carousel" class="right carousel-control"><span class="glyphicon glyphicon-chevron-right"></span></a></div>';
-    
-    $.ajax({
-         url: url,
-         dataType: 'html',
-         success: function(data) {
-         var tHtml="";  
-           if(itemListArray!="" && itemListArray!=undefined){ 
-          for(var i=0; i< itemListArray.length; i++){
-                if(is_cover!=true){
-               var listHtml = data.replace("{title}", itemListArray[i].name);
-               listHtml = listHtml.replace("{sub_title}",itemListArray[i].name);
-               listHtml = listHtml.replace("{about}",itemListArray[i].about);
-              // listHtml = listHtml.replace("{image}", itemListArray[i].image);
-              
-               listHtml = listHtml.replace("{image}","http://wrio.s3-website-us-east-1.amazonaws.com/Default-WRIO-Theme/img/no-photo-200x200.png");
-               listHtml = listHtml.replace("{url}", itemListArray[i].url);
-               
-               listHtml = listHtml.replace("{created_date}","22 Jun 2013");
-               listHtml = listHtml.replace("{rating}", "244");
-               listHtml = listHtml.replace("{readers}", "1,634");
-               listHtml = listHtml.replace("{access}", "Free");
-               tHtml=tHtml+listHtml ;
-            
-             }else{  // for cover structure
-                  
-                 var listHtml = coverHtml.replace("{name}", itemListArray[i].name);
-                 var text=""; 
-                 if(itemListArray[i].text != undefined){
-                  
-                   for(var j=0; j< itemListArray[i].text.length; j++){
-                      li=itemListArray[i].text[j];
-                      var allcont = contentdiv.replace("{content}",li);          
-                      text=text + allcont; 
-                   }  
-                   
-                 }
-                  
-                        
-                       listHtml = listHtml.replace("{contentUrl}",itemListArray[i].contentUrl);
-                listHtml = listHtml.replace("{text}",text); 
-
-                 if(i==0){
-                   listHtml = listHtml.replace("{status}",'item active'); 
-                 }else{
-                    listHtml = listHtml.replace("{status}",'item');
-                 }
-                 tHtml=tHtml+listHtml;  
-                
-              }
-          }
-           }
-         
-        if(is_cover==true){
-            tHtml=coverHeader+tHtml+coverFooter;
-            is_cover=false;
-        }
-         
-            is_append=false;
-            if(tHtml!="") {  
-          $('.itemListWrp').css('display','block');
-             if($(".itemListWrp").length==0){
-                tHtml='<div class="itemListWrp">'+tHtml+'</div>';
-                is_append=true;
-             }
-             
-             if(is_airticlelist==false){
-                $("section").parent("article").css('display','block');
-                          $('#titter_frame_container').css('display','block');
-            }else{
-                 $("section").parent("article").css('display','none');
-                           $('#titter_frame_container').css('display','none');
-             }
-          }
-          
-          if($(".paragraph").length==0){ 
-                if(is_append==true){
-                $('#centerWrp').append(tHtml);
-                $('#titter_frame_container').remove();
-              $('#titter-id').remove();
-                    $('#titter-template').remove();
-              }else{
-                $('.itemListWrp').html(tHtml);
-                $('#titter_frame_container').remove();
-              $('#titter-id').remove();
-                    $('#titter-template').remove();
-              } 
-          }else{
-                  if(is_append==true){
-                   $('#centerWrp').append(tHtml);
-              }else{
-                   $('.itemListWrp').html(tHtml);
-              }  
-          }                                       
-          
-        }
-     });
- }
- // for list  
-
-
-// function for check url
-function checkUrl(){
-     if(href!= undefined ){
-       href = href.replace('index.html', '');  // for remove index.html
-       href = href.replace('index.htm', '');  // for remove index.htm
-     }
-   if(href!= undefined && href.substr(-1) == '/') {
-    href  = href.substring(0,href.length - 1);   // for remove "/" from string "30-04-15"
-   }
-       //console.log(finalMenuJsonArray);
-    
-     // for remove # from url
-     hashName='';
-   if(href.indexOf("#") != -1){
-     var hrefArray = href.split("#"); 
-     url=hrefArray['0']?hrefArray['0']:'';
-     hashName=hrefArray['1']?hrefArray['1']:'';
-       is_hashUrl=true;
-  }
-     hashUrl="";
-     for (i=0;i<finalMenuJsonArray.length;i++){
-                if(finalMenuJsonArray[i].name==hashName){
-              hashUrl=finalMenuJsonArray[i].url;
-             getListJson(hashUrl); 
-          }     
-     }  
-        
-}
-// end fn
-
 domready(function () {
-	React.render(
-		<Main />,
-		document.body
-	);
+    React.render(
+        <Main />,
+        document.body
+    );
 });
